Put duplicate code in its own class
